+++ src/main/java/com/ll/gooHaeYu/domain/chat/message/controller/MessageController.java
@@ -4,7 +4,7 @@ import com.ll.gooHaeYu.domain.chat.message.dto.MessageDto;
 import com.ll.gooHaeYu.domain.chat.message.dto.MessageForm;
 import com.ll.gooHaeYu.domain.chat.message.entity.Message;
 import com.ll.gooHaeYu.domain.chat.message.service.MessageService;
-import com.ll.gooHaeYu.global.rsData.ApiResponse;
+import com.ll.gooHaeYu.global.apiResponse.ApiResponse;
 import com.ll.gooHaeYu.global.security.MemberDetails;
 import com.ll.gooHaeYu.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
