+++ src/main/java/com/ll/gooHaeYu/domain/jobPost/employ/controller/EmployController.java
@@ -2,7 +2,7 @@ package com.ll.gooHaeYu.domain.jobPost.employ.controller;
 
 import com.ll.gooHaeYu.domain.application.application.dto.ApplicationDto;
 import com.ll.gooHaeYu.domain.jobPost.employ.service.EmployService;
-import com.ll.gooHaeYu.global.rsData.ApiResponse;
+import com.ll.gooHaeYu.global.apiResponse.ApiResponse;
 import com.ll.gooHaeYu.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
