+++ src/main/java/com/ll/gooHaeYu/domain/jobPost/employ/controller/WorkCompletionController.java
@@ -1,7 +1,7 @@
 package com.ll.gooHaeYu.domain.jobPost.employ.controller;
 
 import com.ll.gooHaeYu.domain.jobPost.employ.service.WorkCompletionService;
-import com.ll.gooHaeYu.global.rsData.ApiResponse;
+import com.ll.gooHaeYu.global.apiResponse.ApiResponse;
 import com.ll.gooHaeYu.global.security.MemberDetails;
 import com.ll.gooHaeYu.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
